Deploy edge case fix. Possible com fix (now only fortify 1 troop)

+++ app/src/main/java/com/example/riskgame/Risk/views/RiskMapView.java
@@ -45,8 +45,8 @@ public class RiskMapView extends FlashSurfaceView {
     private int top = getTop();
     private int right = getRight();
     private int bottom = getBottom();
-    private float shiftX = 0;
-    private float shiftY = 0;
+    private int shiftX = 0;
+    private int shiftY = 0;
 
     /**
      * MapView
